requirements.txt
Flask==0.10.1
Flask-SQLAlchemy==1.0
Jinja2==2.7
MarkupSafe==0.18
MySQL-python==1.2.4
SQLAlchemy==0.8.2
Werkzeug==0.9.3
itsdangerous==0.22
libtaxii==1.0.105
lxml==3.2.3
python-dateutil==2.1
simplejson==3.3.0
six==1.3.0
wsgiref==0.1.2
